Fixed Case.tsx & csamReportApi tests tests. Move newContactState to it's own file
