+++ plugin-hrm-form/src/states/contacts/reducer.ts
@@ -19,18 +19,14 @@ import { callTypes } from 'hrm-form-definitions';
 import { createReducer } from 'redux-promise-middleware-actions';
 
 import * as t from './types';
-import { ContactMetadata, ContactsState } from './types';
+import { ContactsState } from './types';
 import {
-  DefinitionVersion,
   INITIALIZE_CONTACT_STATE,
   InitializeContactStateAction,
   REMOVE_CONTACT_STATE,
   RemoveContactStateAction,
 } from '../types';
-import { createStateItem, getInitialValue } from '../../components/common/forms/formGenerators';
-import { createContactlessTaskTabDefinition } from '../../components/tabbedForms/ContactlessTaskTabDefinition';
 import {
-  ContactState,
   createDraftReducer,
   EXISTING_CONTACT_CREATE_DRAFT_ACTION,
   EXISTING_CONTACT_LOAD_TRANSCRIPT,
@@ -54,93 +50,13 @@ import {
   TOGGLE_DETAIL_EXPANDED_ACTION,
 } from './contactDetails';
 import { ADD_EXTERNAL_REPORT_ENTRY, addExternalReportEntryReducer } from '../csam-report/existingContactExternalReport';
-import { ReferralLookupStatus, resourceReferralReducer } from './resourceReferral';
-import { Contact, ContactRawJson } from '../../types/types';
+import { resourceReferralReducer } from './resourceReferral';
 import { ContactCategoryAction, toggleSubCategoriesReducer } from './categories';
 import { configurationBase, RootState } from '..';
 import { createCaseAsyncAction } from '../case/saveCase';
+import { newContactState } from './contactState';
 
 export const emptyCategories = [];
-
-export const newContactMetaData = (recreated: boolean): ContactMetadata => {
-  const categoriesMeta = {
-    gridView: false,
-    expanded: {},
-  };
-
-  return {
-    draft: {
-      resourceReferralList: {
-        resourceReferralIdToAdd: '',
-        lookupStatus: ReferralLookupStatus.NOT_STARTED,
-      },
-    },
-    startMillis: recreated ? null : new Date().getTime(),
-    endMillis: null,
-    recreated,
-    categories: categoriesMeta,
-  };
-};
-
-const newContact = (definitions: DefinitionVersion): Contact => {
-  const initialChildInformation = definitions.tabbedForms.ChildInformationTab.reduce(createStateItem, {});
-  const initialCallerInformation = definitions.tabbedForms.CallerInformationTab.reduce(createStateItem, {});
-  const initialCaseInformation = definitions.tabbedForms.CaseInformationTab.reduce(createStateItem, {});
-
-  const { helplines } = definitions.helplineInformation;
-  const defaultHelpline = helplines.find(helpline => helpline.default).value || helplines[0].value;
-  if (defaultHelpline === null || defaultHelpline === undefined) throw new Error('No helpline definition was found');
-
-  const initialContactlessTaskTabDefinition = createContactlessTaskTabDefinition({
-    counselorsList: [],
-    definition: definitions.tabbedForms.ContactlessTaskTab,
-    helplineInformation: definitions.helplineInformation,
-  });
-  const contactlessTask: ContactRawJson['contactlessTask'] = {
-    channel: 'web', // default, should be overwritten
-    date: new Date().toISOString(),
-    time: new Date().toTimeString(),
-    createdOnBehalfOf: '',
-    ...Object.fromEntries(initialContactlessTaskTabDefinition.map(d => [d.name, getInitialValue(d)])),
-  };
-
-  return {
-    accountSid: '',
-    id: '',
-    twilioWorkerId: '',
-    timeOfContact: new Date().toISOString(),
-    taskId: '',
-    helpline: '',
-    rawJson: {
-      childInformation: initialChildInformation,
-      callerInformation: initialCallerInformation,
-      caseInformation: initialCaseInformation,
-      callType: '',
-      contactlessTask,
-      categories: {},
-    },
-    createdBy: '',
-    createdAt: '',
-    updatedBy: '',
-    updatedAt: '',
-    queueName: '',
-    channel: 'web',
-    number: '',
-    conversationDuration: 0,
-    channelSid: '',
-    serviceSid: '',
-    csamReports: [],
-    conversationMedia: [],
-  };
-};
-
-// eslint-disable-next-line import/no-unused-modules
-export const newContactState = (definitions: DefinitionVersion) => (recreated: boolean): ContactState => ({
-  savedContact: newContact(definitions),
-  metadata: newContactMetaData(recreated),
-  draftContact: {},
-  references: new Set(),
-});
 
 // exposed for testing
 export const initialState: ContactsState = {
